Move currencyh-related utilities to a new file - currency.py
